backPlacePass/requirements.txt
Django==2.2
django-cors-headers==2.5.2
djangorestframework==3.9.2
djangorestframework-simplejwt==4.3.0
rest-social-auth==2.1.0
django-sslserver==0.20
